#3  newPromotion terminer , ajout des creneaux

+++ src/edt/Classe/Etudiant.java
@@ -23,11 +23,11 @@ public class Etudiant implements Model_JDBC  {
     private String email;
     private String telephone;
     private boolean notificationsActives;
-    private int idPromotion;
+    private long idPromotion;
     private int idSpecialite;
     
     //c'est plus simple pour moi de me promener avec les int promo et int spe que avec les objets
-    public Etudiant(int num, String prenom, String nom, String email, String telephone, boolean notif, int promo, int spe){
+    public Etudiant(int num, String prenom, String nom, String email, String telephone, boolean notif, long promo, int spe){
         this.id = 0;
         this.numeroEtudiant = num;
         this.prenom = prenom;
@@ -65,7 +65,7 @@ public class Etudiant implements Model_JDBC  {
         return id;
     }
 
-    public int getIdPromotion() {
+    public long getIdPromotion() {
         return idPromotion;
     }
